Add fully functioning jQuery/JS code
This file contains the logic behind all functionalities of the
application. This includes HTML injection, where the HTML divs
are populated via jQuery, and object-oriented programming to
handle many use cases.

It is responsible for loading in all Tweets, allowing the user to
click on usernames and buttons, and various animations. It also
handles error-checking and form validation when the user posts a Tweet.

Pushing full file to personal repo.